Moved controllers and clients into one folder
